+++ campus-web/src/views/main/index.vue
@@ -1,12 +1,17 @@
 <template>
-    <main class="main-container">
-        <div class="main">
-            <div class="left-container" :class="{ 'left-top': isHidden }">
-                <leftCpn :leftTabsList="leftTabsList" />
+    <div>
+        <main class="main-container">
+            <div class="main">
+                <div class="left-container" :class="{ 'left-top': isHidden }">
+                    <left-cpn :leftTabsList="leftTabsList" />
+                </div>
+                <main-cpn />
+                <div class="right-container" :class="{ 'right-top': isHidden }">
+                    <right-cpn />
+                </div>
             </div>
-            <right-cpn />
-        </div>
-    </main>
+        </main>
+    </div>
 </template>
 
 <script setup lang="ts">
@@ -14,6 +19,7 @@ import { reactive, ref, watch } from 'vue'
 
 
 import leftCpn from './leftCpn/index.vue'
+import mainCpn from './mainCpn/index.vue'
 import rightCpn from './rightCpn/index.vue'
 
 const { isHidden } = defineProps({
@@ -39,29 +45,34 @@ const leftTabsList = reactive([
         margin-top: 8px;
         display: flex;
         flex-direction: row;
-
         margin-top: 18px;
+        justify-content: space-between;
 
-        .left-top.left-container {
+        .left-top.left-container,
+        .right-top.right-container {
             top: 20px;
             max-height: calc(100vh - 40px);
         }
 
-        &>.left-container {
+        &>.left-container,
+        &>.right-container {
             position: sticky;
             top: 80px;
         }
 
-        .left-container {
+        .left-container,
+        .right-container {
             width: 180px;
-            margin-right: 20px;
+            // margin-right: 20px;
             height: fit-content;
             border-radius: 4px;
             background-color: #fff;
             max-height: calc(100vh - 101px);
             overflow-x: hidden;
+        }
 
-
+        .right-container {
+            width: 260px;
         }
     }
