+++ campus-web/src/views/main/rightCpn/index.vue
@@ -1,90 +1,85 @@
 <template>
-    <div class="right-container">
-        <div class="main-content">
-            <header-cpn />
-            <entry-list :essay="essay" />
-        </div>
-        <aside-cpn />
+    <div>
+        <aside class="aside-wrap">
+            <div class="aside-main">
+                <div class="aside-header">
+                    <div class="title">
+                        热搜榜单
+                    </div>
+                </div>
+                <div class="aside-content">
+                    <div class="main-item" v-for="i of 6" :key="i">
+                        <a href="#">
+                            <div class="item-box">
+                                <div class="title">
+                                    {{ `热搜事件名字${i}` }}
+                                </div>
+                            </div>
+                        </a>
+                    </div>
+                </div>
+            </div>
+        </aside>
     </div>
-    <!-- <header-cpn />
-    <entry-list :essay="essay" /> -->
 </template>
 
 <script setup lang="ts">
 import { reactive } from "vue";
-import headerCpn from './cpn/headerCpn.vue'
-import entryList from './cpn/entryList.vue'
-import asideCpn from './cpn/asideCpn.vue'
 
-const essay = reactive([
-    {
-        title: '集大校园1',
-        text: '正文部分正文部分正文部分正文部分正文部分正文部分正文部分正文部分正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分',
-        id: 1
-    },
-    {
-        title: '集大校园222',
-        text: '正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分',
-        id: 2
-    }, {
-        title: '集大校园3333',
-        text: '正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分 正文部分',
-        id: 3
-    },
-    {
-        title: '测试',
-        text: '测试',
-        id: 4
-    },
-    {
-        title: '测试',
-        text: '测试',
-        id: 5
-    },
-    {
-        title: '测试',
-        text: '测试',
-        id: 6
-    },
-    {
-        title: '测试',
-        text: '测试',
-        id: 7
-    },
-    {
-        title: '测试',
-        text: '测试',
-        id: 8
-    },
-    {
-        title: '测试',
-        text: '测试',
-        id: 8
-    },
-    {
-        title: '测试',
-        text: '测试',
-        id: 9
-    },
-    {
-        title: '测试',
-        text: '测试',
-        id: 10
-    },
-])
 </script>
 
 <style scoped lang="less">
-.right-container {
-    // background-color: #fff;
-    border-radius: 4px;
+.aside-wrap {
+    width: 200px;
+    display: block;
+    background-color: #fff;
 
-    .main-content {
-        margin-right: 240px;
-        border-radius: 2px;
-        width: 720px;
-        position: relative;
-        background-color: #fff;
+    .aside-main {
+        .aside-header {
+            height: 25px;
+            padding: 12px 14px 8px;
+            display: flex;
+            align-items: center;
+
+            .title {
+                align-self: center;
+                color: #333;
+                font-weight: 500;
+                font-size: 16px;
+                line-height: 22px;
+                box-sizing: border-box;
+                flex: 1;
+            }
+        }
+
+        .aside-content {
+            .main-item {
+                width: 100%;
+                padding: 0 14px;
+
+                .item-box {
+                    position: relative;
+                    height: 40px;
+                    padding: 0 26px 0 21px;
+                    overflow: hidden;
+                    white-space: nowrap;
+                    display: flex;
+                    align-items: center;
+
+                    .title {
+                        margin-right: 10px;
+                        color: #333;
+                        font-size: 14px;
+                        line-height: 18px;
+                        overflow: hidden;
+                        text-overflow: ellipsis;
+                        white-space: nowrap;
+                    }
+                }
+            }
+        }
     }
+
+
 }
 </style>
